pyproject.toml
[project]
name = "user-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "behave>=1.3.3",
    "black>=25.12.0",
    "compare>=0.2b0",
    "coverage>=7.13.1",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "httpie>=3.2.4",
    "httpx>=0.28.1",
    "nose>=1.3.7",
    "pinocchio>=0.4.3",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "ruff>=0.15.0",
    "selenium>=4.39.0",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
